# doc/openapi.yaml
openapi: 3.0.3
info:
  title: API Feria
  license:
    name: UNLICENSED
    url: https://unlicense.org/
  description: |
    API REST (NestJS + Prisma) para gestión de carreras, estudiantes, docentes, proyectos,
    jurados, criterios, evaluaciones y relaciones proyecto–estudiante.

    Las fechas almacenadas como `Date` en PostgreSQL se envían/reciben como string ISO 8601
    (típicamente `YYYY-MM-DD` para campos solo-fecha).
  version: "1.0.0"
servers:
  - url: http://localhost:3000
    description: Servidor local (variable de entorno `PORT`, por defecto 3000)

# La API no usa autenticación por ahora; declara explícitamente endpoints públicos.
security: []
...
paths:
...
  /estudiante/{id}:
    get:
      tags: [Estudiante]
      summary: Obtener estudiante por id
      operationId: estudianteFindOne
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                type: object
                nullable: true
                allOf:
                  - $ref: "#/components/schemas/Estudiante"
    patch:
      tags: [Estudiante]
      summary: Actualizar estudiante
      operationId: estudianteUpdate
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEstudiante"
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estudiante"
    delete:
      tags: [Estudiante]
      summary: Eliminar estudiante
      operationId: estudianteRemove
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estudiante"
...
  /estudiante/{idEstudiante}/proyecto/{idProyecto}:
    post:
      tags: [Estudiante]
      summary: Asignar estudiante a proyecto
      operationId: estudianteAsignarProyecto
      parameters:
        - name: idEstudiante
          in: path
          required: true
          schema:
            type: integer
        - name: idProyecto
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Relación creada (forma según servicio)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProyectoEstudiante"
    delete:
      tags: [Estudiante]
      summary: Quitar estudiante del proyecto
      operationId: estudianteRemoverProyecto
      parameters:
        - name: idEstudiante
          in: path
          required: true
          schema:
            type: integer
        - name: idProyecto
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProyectoEstudiante"
...
  /proyecto/{id}:
    get:
      tags: [Proyecto]
      summary: Obtener proyecto por id
      operationId: proyectoFindOne
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                type: object
                nullable: true
                allOf:
                  - $ref: "#/components/schemas/Proyecto"
    patch:
      tags: [Proyecto]
      summary: Actualizar proyecto
      operationId: proyectoUpdate
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProyecto"
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proyecto"
    delete:
      tags: [Proyecto]
      summary: Eliminar proyecto
      operationId: proyectoRemove
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proyecto"
...
  /proyecto/{idProyecto}/estudiante/{idEstudiante}:
    post:
      tags: [Proyecto]
      summary: Asignar estudiante al proyecto
      operationId: proyectoAsignarEstudiante
      parameters:
        - name: idProyecto
          in: path
          required: true
          schema:
            type: integer
        - name: idEstudiante
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Recurso creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProyectoEstudiante"
    delete:
      tags: [Proyecto]
      summary: Quitar estudiante del proyecto
      operationId: proyectoRemoverEstudiante
      parameters:
        - name: idProyecto
          in: path
          required: true
          schema:
            type: integer
        - name: idEstudiante
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProyectoEstudiante"
  /proyecto-estudiante:
    get:
      tags: [ProyectoEstudiante]
      summary: Listar todas las relaciones proyecto–estudiante
      operationId: proyectoEstudianteFindAll
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProyectoEstudiante"
    post:
      tags: [ProyectoEstudiante]
      summary: Crear relación proyecto–estudiante
      operationId: proyectoEstudianteCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProyectoEstudiante"
      responses:
        "201":
          description: Recurso creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProyectoEstudiante"
  /proyecto-estudiante/proyecto/{idProyecto}:
    get:
      tags: [ProyectoEstudiante]
      summary: Relaciones por proyecto
      operationId: proyectoEstudianteByProyecto
      parameters:
        - name: idProyecto
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProyectoEstudiante"
  /proyecto-estudiante/estudiante/{idEstudiante}:
    get:
      tags: [ProyectoEstudiante]
      summary: Relaciones por estudiante
      operationId: proyectoEstudianteByEstudiante
      parameters:
        - name: idEstudiante
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProyectoEstudiante"
  /proyecto-estudiante/proyecto/{idProyecto}/estudiante/{idEstudiante}:
    get:
      tags: [ProyectoEstudiante]
      summary: Obtener una relación concreta
      operationId: proyectoEstudianteFindOne
      parameters:
        - name: idProyecto
          in: path
          required: true
          schema:
            type: integer
        - name: idEstudiante
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProyectoEstudiante"
        "404":
          description: No encontrada
    delete:
      tags: [ProyectoEstudiante]
      summary: Eliminar relación
      operationId: proyectoEstudianteRemove
      parameters:
        - name: idProyecto
          in: path
          required: true
          schema:
            type: integer
        - name: idEstudiante
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProyectoEstudiante"
...
  /evaluacion/{id}:
    get:
      tags: [Evaluacion]
      summary: Obtener evaluación por id
      operationId: evaluacionFindOne
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                type: object
                nullable: true
                allOf:
                  - $ref: "#/components/schemas/Evaluacion"
    patch:
      tags: [Evaluacion]
      summary: Actualizar evaluación (opcionalmente reemplazar detalles)
      operationId: evaluacionUpdate
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEvaluacion"
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evaluacion"
    delete:
      tags: [Evaluacion]
      summary: Eliminar evaluación
      operationId: evaluacionRemove
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evaluacion"
...
  /detalle-evaluacion:
    get:
      tags: [DetalleEvaluacion]
      summary: Listar detalles de evaluación
      operationId: detalleEvaluacionFindAll
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DetalleEvaluacion"
    post:
      tags: [DetalleEvaluacion]
      summary: Crear detalle (recalcula puntaje total de la evaluación)
      operationId: detalleEvaluacionCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDetalleEvaluacion"
      responses:
        "201":
          description: Recurso creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetalleEvaluacion"
  /detalle-evaluacion/evaluacion/{idEvaluacion}:
    get:
      tags: [DetalleEvaluacion]
      summary: Detalles por id de evaluación
      operationId: detalleEvaluacionByEvaluacion
      parameters:
        - name: idEvaluacion
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DetalleEvaluacion"
  /detalle-evaluacion/{id}:
    get:
      tags: [DetalleEvaluacion]
      summary: Obtener detalle por id
      operationId: detalleEvaluacionFindOne
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetalleEvaluacion"
        "404":
          description: No encontrado
    patch:
      tags: [DetalleEvaluacion]
      summary: Actualizar detalle (recalcula puntaje total)
      operationId: detalleEvaluacionUpdate
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDetalleEvaluacion"
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetalleEvaluacion"
    delete:
      tags: [DetalleEvaluacion]
      summary: Eliminar detalle (recalcula puntaje total)
      operationId: detalleEvaluacionRemove
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Respuesta correcta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetalleEvaluacion"
components:
  parameters:
    PathId:
      name: id
      in: path
      required: true
      description: Identificador numérico
      schema:
        type: integer
  schemas:
    Carrera:
      type: object
      properties:
        id_carrera:
          type: integer
        nombre_carrera:
          type: string
        codigo_carrera:
          type: string
        estudiantes:
          type: array
          items:
            $ref: "#/components/schemas/Estudiante"
        proyectos:
          type: array
          items:
            $ref: "#/components/schemas/Proyecto"
...
    Estudiante:
      type: object
      properties:
        id_estudiante:
          type: integer
        ci:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        telefono:
          type: string
          nullable: true
        correo:
          type: string
          nullable: true
        id_carrera:
          type: integer
        carrera:
          $ref: "#/components/schemas/Carrera"
        proyectos:
          type: array
          items:
            $ref: "#/components/schemas/ProyectoEstudiante"
...
    UpdateEstudiante:
      type: object
      properties:
        ci:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        telefono:
          type: string
        correo:
          type: string
        id_carrera:
          type: integer
    Docente:
      type: object
      properties:
        id_docente:
          type: integer
        ci:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        telefono:
          type: string
          nullable: true
        correo:
          type: string
          nullable: true
        especialidad:
          type: string
          nullable: true
        proyectos:
          type: array
          items:
            $ref: "#/components/schemas/Proyecto"
...
    Proyecto:
      type: object
      properties:
        id_proyecto:
          type: integer
        titulo:
          type: string
        descripcion:
          type: string
          nullable: true
        area_tematica:
          type: string
          nullable: true
        fecha_registro:
          type: string
          format: date
        id_docente:
          type: integer
        id_carrera:
          type: integer
        docente:
          $ref: "#/components/schemas/Docente"
        carrera:
          $ref: "#/components/schemas/Carrera"
        estudiantes:
          type: array
          items:
            $ref: "#/components/schemas/ProyectoEstudiante"
        evaluaciones:
          type: array
          items:
            $ref: "#/components/schemas/Evaluacion"
...
    UpdateProyecto:
      type: object
      properties:
        titulo:
          type: string
        descripcion:
          type: string
        area_tematica:
          type: string
        fecha_registro:
          type: string
          format: date
        id_docente:
          type: integer
        id_carrera:
          type: integer
    ProyectoEstudiante:
      type: object
      properties:
        id_proyecto:
          type: integer
        id_estudiante:
          type: integer
        proyecto:
          $ref: "#/components/schemas/Proyecto"
        estudiante:
          $ref: "#/components/schemas/Estudiante"
    CreateProyectoEstudiante:
      type: object
      required: [id_proyecto, id_estudiante]
      properties:
        id_proyecto:
          type: integer
        id_estudiante:
          type: integer
    Jurado:
      type: object
      properties:
        id_jurado:
          type: integer
        ci:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        institucion:
          type: string
          nullable: true
        especialidad:
          type: string
          nullable: true
        telefono:
          type: string
          nullable: true
        correo:
          type: string
          nullable: true
        evaluaciones:
          type: array
          items:
            $ref: "#/components/schemas/Evaluacion"
...
    Criterio:
      type: object
      properties:
        id_criterio:
          type: integer
        nombre_criterio:
          type: string
        descripcion:
          type: string
          nullable: true
        puntaje_maximo:
          type: number
          format: float
...
    Evaluacion:
      type: object
      properties:
        id_evaluacion:
          type: integer
        id_jurado:
          type: integer
        id_proyecto:
          type: integer
        fecha_evaluacion:
          type: string
          format: date
        observaciones:
          type: string
          nullable: true
        puntaje_total:
          type: number
          format: float
          nullable: true
        jurado:
          $ref: "#/components/schemas/Jurado"
        proyecto:
          $ref: "#/components/schemas/Proyecto"
        detalles:
          type: array
          items:
            $ref: "#/components/schemas/DetalleEvaluacion"
    EvaluacionDetalleItem:
      type: object
      required: [id_criterio, puntaje_obtenido]
      properties:
        id_criterio:
          type: integer
        puntaje_obtenido:
          type: number
          format: float
...
    UpdateEvaluacion:
      type: object
      properties:
        id_jurado:
          type: integer
        id_proyecto:
          type: integer
        fecha_evaluacion:
          type: string
          format: date
        observaciones:
          type: string
        detalles:
          type: array
          items:
            $ref: "#/components/schemas/EvaluacionDetalleItem"
    DetalleEvaluacion:
      type: object
      properties:
        id_detalle:
          type: integer
        id_evaluacion:
          type: integer
        id_criterio:
          type: integer
        puntaje_obtenido:
          type: number
          format: float
        evaluacion:
          $ref: "#/components/schemas/Evaluacion"
        criterio:
          $ref: "#/components/schemas/Criterio"
    CreateDetalleEvaluacion:
      type: object
      required: [id_evaluacion, id_criterio, puntaje_obtenido]
      properties:
        id_evaluacion:
          type: integer
        id_criterio:
          type: integer
        puntaje_obtenido:
          type: number
          format: float
    UpdateDetalleEvaluacion:
      type: object
      properties:
        id_evaluacion:
          type: integer
        id_criterio:
          type: integer
        puntaje_obtenido:
          type: number
          format: float